fix: run manager tests in BeforeSuite
